refactor(domain): extract dashboard html rendering module
